Hoist randomization of arguments to crossover
There is a subtle bug that occurs when the first genome argument supplied to a crossover algorithm is always used for the start of the resulting chromosome.
This leads to a rapid and fatal loss of diversity.
The fix is to randomly swap the two genome arguments to the crossover algorithm.
This commit moves that code to into the solve method.
The result is that all crossover implementors are protected from the bug and resulting implementations are smaller and easier to write.

This commit also removes an unnecessary clone of each chromosome.